Updating replication configuration for Percona 5.5
Using the master-* variables to set up replication in my.cnf has been
deprecated in MySQL 5.1. Replication is now set up via the CHANGE MASTER
TO statement. An extra recipe for replication has been created to
accomodate this.

Also, some default attributes have been changed to match the default
settings for Percona 5.5. Using anything other than the default settings
for innodb_buffer_pool_size and innodb_log_*_size will make this
cookbook fail (the daemon starts with the default configuration at
package install, and when restarting after reconfiguration it will fail
to start because the InnoDB config does not match). There might be a
better way around this, but I have not found one.

With these changes, replication should work out of the box. Just install
a master and a slave with the cookbook and run "START SLAVE;" on the
slave machine.