Update Percona Signing Key
On December 28th 2018, Percona changed their signing key for new packages.
This means packages cannot be downloaded via Yum unless the key URL is updated.

There are still packages available on the repository that are signed with the old key,
so we need to have both keys available.

See https://www.percona.com/blog/2019/02/05/new-percona-package-signing-key-requires-update-on-rhel-and-centos/